Module3_Predictive_Analytics
Summary:
This project is related to one of the leading telecom company. They hired our services to deal with the customer attrition challenge. They had planned to conduct certain promotional and retention campaigns for customers, who were likely to leave their services. The primary objective of this collaboration was to identify customers in risk bucket, who are highly likely to switch to some other service provider.